+++ example/anomaly-monitor/Env.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE TemplateHaskell #-}
+
 module Env
   ( Env (..)
   , HasLogAction (..)