+++ src/Himari/Env/Simple.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE Trustworthy #-}
 
 -- | 基本的な環境を提供するモジュール。