+++ src/Himari/Prelude/Arrow.hs
@@ -1,7 +1,5 @@
 {-# LANGUAGE Safe #-}
-{-# LANGUAGE NoDerivingVia #-}
 {-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
-{-# LANGUAGE NoTemplateHaskell #-}
 
 -- | "Control.Arrow" re-exports, hiding symbols that conflict with "Data.Bifunctor".
 module Himari.Prelude.Arrow