+++ src/Himari/Prelude/Category.hs
@@ -1,7 +1,5 @@
 {-# LANGUAGE Safe #-}
-{-# LANGUAGE NoDerivingVia #-}
 {-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
-{-# LANGUAGE NoTemplateHaskell #-}
 
 -- | "Control.Category" re-exports, hiding symbols that conflict with "Prelude".
 module Himari.Prelude.Category